% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries}
\alias{boundaries}
\title{Boundaries dataset provides the reference movement_ID, ward name, geometry of the region of Bengaluru}
\format{
A data frame with 198 rows and 3 variables:
\describe{
\item{WARD_NAME}{198 Ward names of Bangalore are given}
\item{MOVEMENT_ID}{Unique ID of each and every region in Bangalore}
\item{geometry}{Multipolygon sf objects for each of the regions}
}
}
\source{
{Created using the uber movement dataset that's freely available.}
}
\usage{
boundaries
}
\description{
Contains three variables of importance
}
\examples{
data(boundaries)    #Loads the dataset

}
\keyword{datasets}
