% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_hour}
\alias{date_hour}
\title{Data is to sourced from Uber raw dataset and with modifications, we have it here}
\format{
A data frame with 7648077 rows and 6 variables:
\describe{
\item{sourceid}{refers to one of the 198 regions of Bengaluru region, can also be called as pickup area}
\item{dstid}{refers to one of the 198 regions of Bengaluru region, can also be called as drop area}
\item{month}{refers to the month of the year}
\item{day}{refers to the day of the month}
\item{start_hour}{refers to the start hour in a 24 hour clock}
\item{end_hour}{refers to the end hour in a 24 hour clock}
}
}
\source{
{Created using the uber movement dataset that's freely available.}
}
\usage{
date_hour
}
\description{
Contains 6 variables of importance for our ggplot2 visualisations
}
\examples{
data(date_hour)     #Loads the dataset

}
\keyword{datasets}
